backend/requirements.txt
fastapi
uvicorn
pandas
sqlalchemy
gspread
google-auth
python-multipart
requests
